Trivial fix for statement_mem check overflow.

+++ src/backend/utils/resscheduler/resqueue.c
@@ -2173,7 +2173,7 @@ uint64 ResourceQueueGetQueryMemoryLimit(PlannedStmt *stmt, Oid queueId)
 	/**
 	 * If user requests more using statement_mem, grant that.
 	 */
-	if (queryMem < statement_mem * 1024L)
+	if (queryMem < (uint64) statement_mem * 1024L)
 	{
 		queryMem = (uint64) statement_mem * 1024L;
 	}
